Feature/project aula08

+++ aula-08/jpeg/java/src/main/java/com/github/newtonjose/cs/aula08/jpeg/src/VerificaJpegBytes.java
@@ -61,7 +61,7 @@ public class VerificaJpegBytes {
      */
     private byte[] lerDoisUltimosBytes(final int len) throws IOException {
         final byte[] data = new byte[2];
-        randomAccessFile.seek(len - 2);
+        randomAccessFile.seek((long) len - 2);
         randomAccessFile.read(data);
 
         return data;
